Rewrite project using ReactJS
Rewrite the project using ReactJS.

* **Add ReactJS structure:**
  - Add `public/index.html` with a root div and script link.
  - Add `src/index.js` to initialize and render the React app.
  - Add `src/App.js` with a functional component displaying the header.

* **Add React components:**
  - Add `src/components/Activity.js` for displaying activity images.
  - Add `src/components/Controls.js` for speech recognition and next activity buttons.
  - Add `src/components/Feedback.js` for displaying feedback text.
  - Add `src/components/Question.js` for displaying question text.

* **Add utility functions:**
  - Add `src/utils/speechRecognition.js` for handling speech recognition.
  - Add `src/utils/textToSpeech.js` for handling text-to-speech.

* **Add data file:**
  - Add `src/data/activities.js` with an array of activity objects.

* **Add package.json:**
  - Add dependencies for React and ReactDOM.
  - Add scripts for starting and building the React app.

* **Remove old files:**
  - Delete `index.html`, `js/script.js`, `winteractivitygame.html`, and `winteractivitygame2 copy.html`.
