--- index.html
@@ -1,131 +1,0 @@
-<!DOCTYPE html>
-<html lang="de">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Winteraktivitäten Spiel</title>
-  <style>
-    body {
-      font-family: Arial, sans-serif;
-      text-align: center;
-      padding: 20px;
-      background-color: #f0f8ff;
-    }
-    
-    .question {
-      font-size: 24px;
-      margin: 20px 0;
-      color: #2c3e50;
-    }
-    
-    #activityImage {
-      max-width: 500px;
-      border-radius: 10px;
-      box-shadow: 0 4px 8px rgba(0,0,0,0.1);
-      margin: 20px 0;
-    }
-    
-    .button {
-      padding: 10px 20px;
-      font-size: 16px;
-      margin: 10px;
-      border: none;
-      background-color: #3498db;
-      color: white;
-      cursor: pointer;
-      border-radius: 5px;
-      transition: all 0.3s ease;
-    }
-    
-    .button:hover {
-      background-color: #2980b9;
-    }
-    
-    .mic-button {
-      width: 80px;
-      height: 80px;
-      border-radius: 50%;
-      background-color: #e74c3c;
-      border: none;
-      cursor: pointer;
-      margin: 20px;
-      box-shadow: 0 4px 8px rgba(0,0,0,0.2);
-      transition: all 0.3s ease;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      padding: 0;
-      position: relative;
-    }
-    
-    .mic-button::after {
-      content: '';
-      display: block;
-      width: 24px;
-      height: 24px;
-      background-color: white;
-      border-radius: 3px;
-    }
-    
-    .mic-button.recording::after {
-      width: 20px;
-      height: 20px;
-    }
-    
-    .mic-button:hover {
-      background-color: #c0392b;
-      transform: scale(1.05);
-    }
-    
-    .mic-button.recording {
-      animation: pulse 1.5s infinite;
-    }
-    
-    @keyframes pulse {
-      0% {
-        transform: scale(1);
-        box-shadow: 0 4px 8px rgba(0,0,0,0.2);
-      }
-      50% {
-        transform: scale(1.05);
-        box-shadow: 0 8px 16px rgba(0,0,0,0.3);
-      }
-      100% {
-        transform: scale(1);
-        box-shadow: 0 4px 8px rgba(0,0,0,0.2);
-      }
-    }
-    
-    .feedback {
-      margin-top: 20px;
-      font-size: 18px;
-      min-height: 27px;
-    }
-    
-    .controls {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      gap: 20px;
-      margin: 20px 0;
-    }
-  </style>
-</head>
-<body>
-  <h1>Winteraktivitäten Spiel</h1>
-
-  <p id="question" class="question">Was machen diese Leute?</p>
-
-  <!-- Image section -->
-  <img id="activityImage" src="" alt="Winteraktivität Bild">
-
-  <div class="controls">
-    <button class="mic-button" onclick="startSpeechRecognition()" id="micButton"></button>
-    <button class="button" onclick="nextImage()">Nächste Aktivität</button>
-  </div>
-
-  <p id="feedback" class="feedback"></p>
-
-  </body>
-  <script src="js/script.js"></script>
-</html>
