+++ package.json
@@ -1,0 +1,19 @@
+{
+  "name": "winteractivitygame",
+  "version": "1.0.0",
+  "description": "A ReactJS project for the Winteraktivitäten Spiel",
+  "main": "src/index.js",
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "dependencies": {
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
+  },
+  "devDependencies": {},
+  "author": "",
+  "license": "ISC"
+}
