+++ src/App.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const App = () => {
+  return (
+    <div>
+      <h1>Winteraktivitäten Spiel</h1>
+    </div>
+  );
+};
+
+export default App;
