+++ src/components/Activity.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Activity = ({ src }) => {
+  return <img src={src} alt="Winteraktivität Bild" />;
+};
+
+export default Activity;
