+++ src/components/Feedback.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Feedback = ({ text }) => {
+  return <p>{text}</p>;
+};
+
+export default Feedback;
