+++ src/components/Question.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Question = ({ text }) => {
+  return <p>{text}</p>;
+};
+
+export default Question;
